+++ .gitignore
@@ -260,3 +260,4 @@ paket-files/
 __pycache__/
 *.pyc
 source/nuget.config
+/Open.Numeric.Primes.Benchmarks/BenchmarkDotNet.Artifacts/results
