+++ Open.Numeric.Primes.Benchmarks/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using Open.Numeric.Primes.Benchmarks;
+
+BenchmarkRunner.Run<IsPrimeBenchmarks>();
