+++ Open.Numeric.Primes.sln
@@ -1,13 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31410.357
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33110.190
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Open.Numeric.Primes", "source\Open.Numeric.Primes.csproj", "{A1D27164-24BD-4DE5-A783-7BF93B32253B}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Open.Numeric.Primes.Speed", "speed\Open.Numeric.Primes.Speed.csproj", "{12663275-69B7-4B4E-899B-07267780F7F0}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Open.Numeric.Primes.Tests", "tests\Open.Numeric.Primes.Tests.csproj", "{AE71AD73-B001-45E7-BC34-CC220CF29CD6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Open.Numeric.Primes.Benchmarks", "Open.Numeric.Primes.Benchmarks\Open.Numeric.Primes.Benchmarks.csproj", "{35374F1D-1B3F-45CB-903F-524A6191C138}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{AE71AD73-B001-45E7-BC34-CC220CF29CD6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AE71AD73-B001-45E7-BC34-CC220CF29CD6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AE71AD73-B001-45E7-BC34-CC220CF29CD6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35374F1D-1B3F-45CB-903F-524A6191C138}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35374F1D-1B3F-45CB-903F-524A6191C138}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35374F1D-1B3F-45CB-903F-524A6191C138}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35374F1D-1B3F-45CB-903F-524A6191C138}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
